+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+- Index :namespace/docstring, :var/docstring and :function/args
+
 ## 0.2.1 (2019-10-29)
 
 ### Fixes